requirements.txt
brotlipy==0.3.0
cffi==1.7.0
dj-database-url==0.4.1
Django==1.9.7
django-cors-headers==1.1.0
django-extensions==1.6.7
djangorestframework==3.3.3
gunicorn==19.6.0
numpy==1.11.0
oauthlib==0.7.2
Pillow==3.2.0
psycopg2==2.6.1
pycparser==2.14
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
tweepy==3.3.0
virtualenv==13.1.0
whitenoise==3.2
